Reorganizing JS files
