+++ traffic_stops/static/js/base/DataHandlerBase.js
