+++ traffic_stops/static/js/base/VisualBase.js
