+++ traffic_stops/static/js/states/nc/index.js
@@ -1,6 +1,6 @@
-import DataHandlerBase from '../util/DataHandlerBase.js';
-import VisualBase from '../util/VisualBase.js';
-import TableBase from '../util/TableBase.js';
+import DataHandlerBase from '../../base/DataHandlerBase.js';
+import VisualBase from '../../base/VisualBase.js';
+import TableBase from '../../base/TableBase.js';
 
 var _ = require('underscore');
 var d3 = require('d3');
